+++ zswap/src/prove.rs
@@ -245,12 +245,12 @@ mod tests {
         let inp =
             Input::new_contract_owned(&mut rng, &qcoin, None, Default::default(), &tree).unwrap();
         let inp_proven = inp.prove(provider.split()).await.unwrap();
-        assert_eq!(inp_proven.proof.0.len(), INPUT_PROOF_SIZE);
+        assert_eq!(inp_proven.proof.bytes.len(), INPUT_PROOF_SIZE);
 
         let out =
             Output::<_, InMemoryDB>::new_contract_owned(&mut rng, &coin, None, Default::default())
                 .unwrap();
         let out_proven = out.prove(provider).await.unwrap();
-        assert_eq!(out_proven.proof.0.len(), OUTPUT_PROOF_SIZE);
+        assert_eq!(out_proven.proof.bytes.len(), OUTPUT_PROOF_SIZE);
     }
 }